backend/app/requirements.txt
fastapi>=0.128.3
uvicorn>=0.40.0
python-jose[cryptography]>=3.5.0
argon2-cffi>=25.1.0
python-multipart>=0.0.22
sqlalchemy>=2.0.46
pymysql>=1.1.2
python-dotenv>=1.2.1
